+++ CHANGES.md
@@ -7,3 +7,9 @@
 - add voice to UserNotification
 - unicode for models represantation
 - added setup to urlpatterns
+
+
+# dimosy django 2016-01-11
+
+- fixed issue #17
+- added "active" to time window class